requirements.txt
matplotlib==3.1.2
more-itertools==9.0.0
NBT==1.5.0
numpy==1.19.3
opencv_python==4.5.1.48
PyGLM==2.6.0
pyglm-typing==0.2.1
requests==2.22.0
scipy==1.9.3
